Update patch set 2
Patch Set 2: Code-Review-1

I'd rather we didn't change the default, as once distil v1 is deprecated (and that will be soon), this makes less sense.

I'd rather than in the places for our own deployment where we are using the distilclient we pass "ratingv2" as a param. That's the reason for having that param, and by doing it that way we keep upstream code cleaner.

For distl-ui we can even make it a config option. If 'DISTIL_SERVICE_TYPE' is defined, use that when making the client, else use default. We can deprecate and remove that setting from distil-ui later as well once v1 is entirely out of the picture.

Patch-set: 2
Reviewer: Gerrit User 10420 <10420@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1